pyproject.toml
[project]
name = "acocalculator"
version = "1.0.0"
description = "Sistema de Gestión de Calificaciones de Moodle - Procesamiento automatizado de trabajos prácticos, parciales y recuperatorios"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sistema ACOCalculator"}
]
keywords = ["moodle", "calificaciones", "educacion", "csv", "notas"]

dependencies = [
    "xlwt==1.3.0",
]

[project.scripts]
acocalculator = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "faker>=20.0.0",
    "factory-boy>=3.3.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Pruebas unitarias",
    "integration: Pruebas de integración",
    "slow: Pruebas que tardan más tiempo",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

